refactor: separate scheduling and task persistence
