+++ backend/repositories/tasks/__init__.py
@@ -1,1 +1,5 @@
 """Task and scheduling persistence repositories."""
+
+from .db import TaskDB
+
+__all__ = ["TaskDB"]
